+++ apps/indexer/config/prod/ganache.exs
@@ -18,7 +18,7 @@ config :indexer,
     transport_options: [
       http: EthereumJSONRPC.HTTP.HTTPoison,
       url: System.get_env("ETHEREUM_JSONRPC_HTTP_URL") || "http://localhost:7545",
-      fallback_url: System.get_env("ETHEREUM_JSONRPC_FALLBACK_HTTP_URL") || "http://localhost:7545",
+      fallback_url: System.get_env("ETHEREUM_JSONRPC_FALLBACK_HTTP_URL"),
       http_options: [recv_timeout: timeout, timeout: timeout, hackney: hackney_opts]
     ],
     variant: EthereumJSONRPC.Ganache